requirements.txt
fastapi
uvicorn
pymupdf
PyPDF2
transformers
pytesseract
Pillow
spacy
googletrans==4.0.0-rc1
textblob
python-docx
